tests/test_lang_detector.py
import unittest
from modules.lang_detector import LanguageDetector

class TestLanguageDetector(unittest.TestCase):
    def test_bad_input(self):
        lang_texts = [
            (None, ''),
            # ('en', '123 456 789.'),
            # ('zh', '*&%544() !@#$^%^'),
        ]
        detector = LanguageDetector()
        for expected_lang, text in lang_texts:
            detected_lang = detector.predict_lang(text)

            self.assertEqual(detected_lang, expected_lang)


    def test_detectLanguages(self):
        # source https://europa.eu/european-union/about-eu/eu-languages_en
        lang_texts = [
            ('bg', 'Европейският съюз е семейство от демократични европейски страни, решени да работят заедно за мир и благоденствие.'),
            ('hr', 'Europska unija (EU) zajednica je europskih demokratskih država koje zajedničkim naporima teže uspostavi mira i blagostanja.'),
            ('cs', 'Evropská unie (EU) je rodina demokratických evropských států, které se zavázaly spolupracovat v zájmu zachování míru a prosperity.'),
            ('da', 'Den Europæiske Union (EU) er en familie af demokratiske lande i Europa, som har forpligtet sig til at arbejde sammen for fred og fremskridt.'),
            ('nl', 'De Europese Unie (EU) is een familie van democratische Europese landen die gezamenlijk aan vrede en welvaart willen werken.'),
            ('en', 'The European Union (EU) is a family of democratic European countries, committed to working together for peace and prosperity.'),
            ('et', 'Euroopa Liit (EL) on demokraatlike Euroopa riikide pere, kes on otsustanud töötada koos rahu ja õitsengu saavutamiseks.'),
            ('fi', 'Euroopan unioni (EU) koostuu demokraattisista valtioista, jotka ovat sitoutuneet työskentelemään yhdessä rauhan ja hyvinvoinnin saavuttamiseksi.'),
            ('fr', 'L’Union européenne (UE) est une famille de pays démocratiques européens décidés à œuvrer ensemble à la paix et à la prospérité.'),
            ('de', 'Die Europäische Union (EU) ist ein Zusammenschluss demokratischer europäischer Länder, die sich der Wahrung des Friedens und dem Streben nach Wohlstand verschrieben haben.'),
            ('el', 'Η Ευρωπαϊκή Ένωση (ΕΕ) είναι μια οικογένεια ευρωπαϊκών δημοκρατικών χωρών, οι οποίες έχουν δεσμευτεί να συνεργάζονται για την ειρήνη και την ευημερία.'),
            ('hu', 'Az Európai Unió (EU) olyan demokratikus európai országok családja, amelyek a békéért és a fejlődésért dolgoznak együtt.'),
            ('ga', 'Cuimsíonn an tAontas Eorpach (AE) grúpa de thíortha Eorpacha daonlathacha atá ag obair as lámha a chéile i dtreo na síochána agus an rathúnais.'),
            ('it', 'L’Unione europea (UE) è una famiglia di paesi europei democratici che si sono impegnati a lavorare insieme per la pace e la prosperità.'),
            ('lv', 'Eiropas Savienība (ES) ir demokrātisku Eiropas valstu apvienība, kuras ietvaros valstis apņēmušās kopīgi strādāt miera un labklājības sasniegšanai.'),
            ('lt', 'Europos Sąjunga (ES) yra demokratinių Europos valstybių šeima, pasiryžusi taikos ir gerovės labui veikti išvien.'),
            ('mt', "L-Unjoni Ewropea (UE) tirrappreżenta familja ta' pajjiżi demokratiċi Ewropej b'impenn li jaħdmu flimkien għall-paċ u l-ġid."),
            ('pl', 'Unia Europejska (UE) to rodzina demokratycznych państw europejskich, których celem jest wspólna praca na rzecz pokoju i dobrobytu.'),
            ('pt', 'A União Europeia (UE) é uma família de países democráticos europeus, com um projecto comum de paz e prosperidade.'),
            ('ro', 'Uniunea Europeană (UE) este o familie de ţări democratice europene, care s-au angajat să lucreze împreună pentru pace şi prosperitate.'),
            ('sk', 'Európska únia (EÚ) je spoločenstvom demokratických európskych štátov, ktoré sa zaviazali spolupracovať na dosiahnutí mieru a prosperity.'),
            ('sl', 'Evropska unija (EU) je družina demokratičnih evropskih držav, ki si skupaj prizadevajo za mir in blaginjo.'),
            ('es', 'La Unión europea (UE) es una familia de países europeos democráticos, que se han comprometido a trabajar juntos en aras de la paz y la prosperidad.'),
            ('sv', 'Europeiska unionen (EU) är en familj av demokratiska europeiska länder som gått samman för att verka för fred och välstånd.'),

            ('ru', 'Русский является также самым распространённым славянским языком[10] и самым распространённым языком в Европе — географически и по числу носителей языка как родного.'),
            ('ja', '使用人口について正確な統計はないが、日本国内の人口、及び日本国外に住む日本人や日系人、日本がかつて統治した地域の一部住民など、約1億3千万人以上と考えられている。'),
            ('zh', '影響包含對氣候的變化以及自然資源的枯竭程度'),
            ('uk', 'У "Газпромі" розповіли, коли буде добудований "Північний потік-2": названі терміни'),
            # be not supported by langdetect
            # ('be', 'Пажар на тарфянішчы: беларускі пратэст мае велізарны патэнцыял'),
        ]
        detector = LanguageDetector()
        for expected_lang, text in lang_texts:
            detected_lang = detector.predict_lang(text)
            if expected_lang in ['nb', 'nn']:
                self.assertIn(detected_lang, [expected_lang, 'no'])
            else:
                self.assertEqual(detected_lang, expected_lang)

    def test_detectMix(self):
        # source https://europa.eu/european-union/about-eu/eu-languages_en
        lang_texts = [
            ('mix1', 'The European Union (EU) is a family of democratic European countries, committed to working together for peace and prosperity. Този текст е на Български.'),
            ('mix2', 'This piece of text is in English. Европейският съюз е семейство от демократични европейски страни, решени да работят заедно за мир и благоденствие.'),
        ]
        detector = LanguageDetector()
        for expected_lang, text in lang_texts:
            detected_lang = detector.predict_lang(text)
            if expected_lang == 'mix1':
                self.assertIn(detected_lang, ['en','bg'])
                # self.assertEqual(detected_lang, 'en') # ftz model fails: en has only 30% vs bg 50% 
            elif expected_lang == 'mix2':
                self.assertIn(detected_lang, ['en','bg'])
                self.assertEqual(detected_lang, 'bg')


    def test_mixWithNewlines(self):
        # source https://europa.eu/european-union/about-eu/eu-languages_en
        lang_texts = [
            ('mix1', 'The European Union (EU) is a family of democratic European countries, \ncommitted to working together for peace and prosperity. \nТози текст е на Български.'),
            ('mix2', 'This piece of text is in English. \nЕвропейският съюз е семейство от демократични европейски страни, \nрешени да работят заедно за мир и благоденствие.'),
        ]
        detector = LanguageDetector()
        for expected_lang, text in lang_texts:
            detected_lang = detector.predict_lang(text)
            if expected_lang == 'mix1':
                self.assertIn(detected_lang, ['en','bg'])
                self.assertEqual(detected_lang, 'en') 
            elif expected_lang == 'mix2':
                self.assertIn(detected_lang, ['en','bg'])
                self.assertEqual(detected_lang, 'bg')



if __name__ == '__main__':
    unittest.main()